Use improved proprietary vendor scripts; add new AR6000 firmware files

+++ copy-files.sh
@@ -20,6 +20,13 @@ DEVICE=galaxy5
 MANUFACTURER=samsung
 
 BASE=../../../vendor/$MANUFACTURER/$DEVICE/proprietary
+if [ "x$1" = "x" ]; then
+  echo "You need to specifify the root of the /system directory"
+  exit
+fi
+SYSDIR=$1
+
+
 rm -rf $BASE/*
 
 for FILE in `cat proprietary-files.txt`; do
@@ -27,7 +34,8 @@ for FILE in `cat proprietary-files.txt`; do
     if [ ! -d $BASE/$DIR ]; then
         mkdir -p $BASE/$DIR
     fi
-    adb pull /system/$FILE $BASE/$FILE
+    cp $SYSDIR/$FILE $BASE/$FILE
 done
 
 ./setup-makefiles.sh
+
